unsure what data type to use, I think this is the best

+++ navigation_unit.h
@@ -16,38 +16,38 @@
 
 struct Com_packet
 {
-    short adress;
-    short packet_count;
-    short data_packets[7];
+    uint8_t adress;
+    uint8_t packet_count;
+    uint8_t data_packets[7];
 };
 
 
-short handle_command(short id);
-short resend(short adress);
-short set_pd_kd(short kd);
-short set_pd_kp(short kp);
-short command_stop();
-short command_start();
-short command_set_target_square(short id);
+uint8_t handle_command(uint8_t id);
+uint8_t resend(uint8_t adress);
+uint8_t set_pd_kd(uint8_t kd);
+uint8_t set_pd_kp(uint8_t kp);
+uint8_t command_stop();
+uint8_t command_start();
+uint8_t command_set_target_square(uint8_t id);
 
-short NAVIGATION_MODE = NAVIGATION_MODE_MANUAL;
-short PD_KD = 0;
-short PD_KP = 0;
+uint8_t NAVIGATION_MODE = NAVIGATION_MODE_MANUAL;
+uint8_t PD_KD = 0;
+uint8_t PD_KP = 0;
 
 // Don't yet know how this is specified for the competition
-int CURRENT_HEADING = FULL_TURN/4;
-short CURRENT_POS_X = 24;
-short CURRENT_POS_Y = 0;
+uint16_t CURRENT_HEADING = FULL_TURN/4;
+uint8_t CURRENT_POS_X = 24;
+uint8_t CURRENT_POS_Y = 0;
 
 // We can either use bool + unsigned
 bool WHEEL_DIR_LEFT = DIR_FORWARD;
 bool WHEEL_DIR_RIGHT = DIR_FORWARD;
-unsigned short WHEEL_SPEED_LEFT = 0;
-unsigned short WHEEL_SPEED_RIGHT = 0;
+uint8_t WHEEL_SPEED_LEFT = 0;
+uint8_t WHEEL_SPEED_RIGHT = 0;
 
 // or a signed value.
-//short WHEEL_SPEED_LEFT = 0;
-//short WHEEL_SPEED_RIGHT = 0;
+//int8_t WHEEL_SPEED_LEFT = 0;
+//int8_t WHEEL_SPEED_RIGHT = 0;
 // depending on what's easiest when translating to PWM values for the robot
 
 // Unsure how to represent this, currently broke it into
@@ -60,10 +60,10 @@ unsigned short WHEEL_SPEED_RIGHT = 0;
 // ignore NAVIGATION_GOAL_HEADING
 #define NAVIGATION_GOAL_MOVE 2
 
-short NAVIGATION_GOAL_TYPE = NAVIGATION_GOAL_NONE;
-short NAVIGATION_GOAL_X = 24;
-short NAVIGATION_GOAL_Y = 0;
-short NAVIGATION_GOAL_HEADING = 0;
+uint8_t NAVIGATION_GOAL_TYPE = NAVIGATION_GOAL_NONE;
+uint8_t NAVIGATION_GOAL_X = 24;
+uint8_t NAVIGATION_GOAL_Y = 0;
+uint8_t NAVIGATION_GOAL_HEADING = 0;
 
 
 #endif
